Agregar, Editar, Eliminar, Crear Productos LISTO!

+++ src/main/java/com/inventory/service/ProductoService.java
@@ -29,33 +29,33 @@ public class ProductoService {
                 throw new RuntimeException("El producto ya existe");
             }
         }
-    
+
         // 🔥 Validar que la categoría existe
         String categoryId = productDto.getCategoryId();
         CategoryProduct categoria = categoryRepository.findById(categoryId)
-            .orElseThrow(() -> new RuntimeException("La categoría con id " + categoryId + " no existe"));
-    
+                .orElseThrow(() -> new RuntimeException("La categoría con id " + categoryId + " no existe"));
+
         // 🔧 Convertir DTO a entidad y setear categoría real
         Product producto = ProductDto.toProducto(productDto);
         producto.setCategory(categoria); // Esta categoría sí existe en la BD
-    
+
         return productRepository.save(producto);
     }
 
-     public List<ProductDto> obtenerProductos() {
+    public List<ProductDto> obtenerProductos() {
         // Obtenemos todos los productos desde la base de datos
         List<Product> productos = productRepository.findAll();
-    
+
         // Convertimos la lista de productos a una lista de ProductDto
         return productos.stream()
-                        .map(ProductDto::new) // Convierte cada producto en ProductDTO
-                        .collect(Collectors.toList());
+                .map(ProductDto::new) // Convierte cada producto en ProductDTO
+                .collect(Collectors.toList());
     }
 
     public Optional<ProductDto> obtenerProductoPorId(String id) {
         // Buscamos el producto por ID
         Optional<Product> producto = productRepository.findById(id);
-        
+
         // Si se encuentra, lo convertimos a ProductDto y lo devolvemos
         return producto.map(ProductDto::new);
     }
@@ -69,10 +69,12 @@ public class ProductoService {
     }
 
     public void eliminarProducto(ProductDto productDto) {
-        // Obtenemos el id del ProductDto
         String id = productDto.getId();
-        
-        // Eliminamos el Producto por id
+
+        if (!productRepository.existsById(id)) {
+            throw new RuntimeException("El producto con id " + id + " no existe");
+        }
+
         productRepository.deleteById(id);
     }
